Changed the lat/long of a parking location to doubles and made the fake data agree with the map data so that navigation actually works.

+++ app/src/main/java/edu/illinois/cs465/parkingpterodactyl/ParkingLocations.java
@@ -6,10 +6,10 @@ package edu.illinois.cs465.parkingpterodactyl;
 
 public class ParkingLocations {
     private String nameOfLoc;
-    private int Latitude;
-    private int Longitude;
+    private double Latitude;
+    private double Longitude;
 
-    public ParkingLocations(String locationName, int LatitudeVal, int LongitudeVal) {
+    public ParkingLocations(String locationName, double LatitudeVal, double LongitudeVal) {
         nameOfLoc = locationName;
         Latitude = LatitudeVal;
         Longitude = LongitudeVal;
@@ -19,11 +19,11 @@ public class ParkingLocations {
         return nameOfLoc;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return Latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return Longitude;
     }
